+++ CHANGELOG.md
@@ -4,6 +4,7 @@ based on [Semantic Versioning](http://semver.org/)
 
 *2.0.0* 2014-03-29
 
++ add a list of identifier attributes in concept
 + add a list of non identifier attributes in concept
 + add missing end of a generated test
 + validate if code is empty in lower or upper transformations in entity